dating-bot/dating_bot/requirements.txt
# Основные зависимости
python-telegram-bot
celery
redis
sqlalchemy
prometheus-client
python-dotenv
